+++ src/main/java/fr/maxlego08/essentials/listener/PlayerListener.java
@@ -7,6 +7,7 @@ import fr.maxlego08.essentials.api.messages.Message;
 import fr.maxlego08.essentials.api.user.Option;
 import fr.maxlego08.essentials.api.user.User;
 import fr.maxlego08.essentials.api.utils.DynamicCooldown;
+import fr.maxlego08.essentials.storage.ConfigStorage;
 import fr.maxlego08.essentials.zutils.utils.TimerBuilder;
 import fr.maxlego08.essentials.zutils.utils.ZUtils;
 import org.bukkit.Material;
@@ -144,6 +145,10 @@ public class PlayerListener extends ZUtils implements Listener {
         User user = this.plugin.getUser(player.getUniqueId());
         if (user != null) user.startCurrentSessionPlayTime();
 
+        if (user != null && user.isFirstJoin() && ConfigStorage.spawnLocation != null) {
+            this.plugin.getScheduler().teleportAsync(player, ConfigStorage.spawnLocation);
+        }
+
         this.plugin.getScheduler().runAtLocationLater(player.getLocation(), () -> {
 
             if (hasPermission(player, Permission.ESSENTIALS_FLY_SAFELOGIN) && shouldFlyBasedOnLocation(player.getLocation())) {
@@ -159,7 +164,7 @@ public class PlayerListener extends ZUtils implements Listener {
     }
 
     @EventHandler(priority = EventPriority.LOWEST)
-    public void onQuid(PlayerQuitEvent event) {
+    public void onQuit(PlayerQuitEvent event) {
         User user = this.plugin.getUser(event.getPlayer().getUniqueId());
         if (user == null) return;
         long sessionPlayTime = (System.currentTimeMillis() - user.getCurrentSessionPlayTime()) / 1000;