+++ changelog.md
@@ -4,6 +4,7 @@
 - Fixed autocompletion for cooldown commands
 - Fixed the cooldown system that could be applied to commands even if an error occurred
 - Fixed folia on player join [#124](https://github.com/Maxlego08/zEssentials/issues/124)
+- Debug player first joins at spawn location [#125](https://github.com/Maxlego08/zEssentials/issues/125)
 
 # 1.0.1.3